# Refactor MiqReport.seed for performance improvements