--- a/spec/models/miq_compare_spec.rb
+++ b/spec/models/miq_compare_spec.rb
@@ -4,7 +4,7 @@ describe MiqCompare do
       vm1 = FactoryBot.create(:vm_vmware)
       vm2 = FactoryBot.create(:vm_vmware)
 
-      MiqReport.seed_report("vms", "compare")
+      MiqReport.seed_report("vms")
 
       report = MiqReport.find_by(:name => "VMs: Compare Template")
       compare = MiqCompare.new({:ids => [vm1.id, vm2.id]}, report)
@@ -19,7 +19,7 @@ describe MiqCompare do
       host2 = FactoryBot.create(:host_vmware)
 
       MiqRegion.seed
-      MiqReport.seed_report("hosts", "compare")
+      MiqReport.seed_report("hosts")
 
       report = MiqReport.find_by(:name => "Hosts: Compare Template")
       compare = MiqCompare.new({:ids => [host1.id, host2.id]}, report)