--- a/spec/models/miq_report/data/product/compare/testing.yaml
+++ b/spec/models/miq_report/data/product/compare/testing.yaml
@@ -1,0 +1,55 @@
+# Report title
+title: "Testing Compare Template"
+
+# Menu name
+menu_name: "Testing Compare Name"
+
+# Main DB table report is based on
+db: VmOrTemplate
+
+# Columns to fetch from the main table
+cols:
+- name
+- vendor
+
+# Included tables (joined, has_one, has_many) and columns
+include:
+
+# Order of columns (from all tables)
+col_order:
+- name
+- vendor
+
+# Column titles, in order
+headers:
+- Name
+- Vendor
+
+# Condition(s) string for the SQL query
+conditions:
+
+# Order string for the SQL query
+order:
+
+# Columns to sort the report on, in order
+sortby:
+
+# Group rows (y=yes,n=no,c=count)
+group:
+
+# Graph type
+#   Bar
+#   Column
+#   ColumnThreed
+#   ParallelThreedColumn
+#   Pie
+#   PieThreed
+#   StackedBar
+#   StackedColumn
+#   StackedThreedColumn
+
+graph:
+
+# Dimensions of graph (1 or 2)
+#   Note: specifying 2 for a single dimension graph may not return expected results
+dims: