Initial seed dropped from 1.27s to 1.21s (4.7% savings - insignificant)
Second seed dropped from 2.20s to 0.06s (97% savings)

@jrafanie Please review
cc @NickLaMuro