@Fryguy 2019-01-07
Also goofed up the numbers on this PR.  I wrote that it went to 0.60s (73%), but it actually went down to 0.06s (97%)!  Updated the commit message and OP accordingly

![mundane_detail](https://user-images.githubusercontent.com/52120/50790057-db450b00-128b-11e9-8135-94b48120e457.gif)