@Fryguy 2019-01-07
I want to write a couple more specs here so please hold on merge