@Fryguy 2019-01-08
good thing I wrote specs...seems I have bugs :)