@Fryguy 2019-01-09
@NickLaMuro @kbrock @jrafanie This should be good to go now.  I added some specs that didn't exist before and did before/after tests to verify.