@Fryguy 2019-01-11
> I didn't see the bug before though.

There was no bug before aside from an N+1 that seemed mostly negligible.  (I wrote specs because I found a bug in my refactoring code - the after code).  So not sure what this comment refers to.  That being said, you could argue the "bug" is that it was slow :smile: