@Fryguy 2019-01-11
> Maybe because it is all in one giant commit so I can't see the logic moving.

Yeah, I was trying to find ways to break this up into commits but honestly, I refactored the logic, so it's more or less a rewrite to get the same results in less time.