@Fryguy 2019-01-15
@jrafanie Updated the specs.