@jrafanie 2019-01-15
I'll merge as is.  The slightly now inaccurate code comment can be corrected in a followup PR.